API

API

Base.Filesystem.isfile — Method.

isfile(rfs::RemoteFileSet, file)

Check whether a specific file contained in rfs has been downloaded.

source

Base.Filesystem.isfile — Method.

isfile(rfs::RemoteFileSet)

Check whether all files contained in rfs have been downloaded.

source

Base.Filesystem.isfile — Method.

isfile(rf::RemoteFile)

Check whether rf has been downloaded.

source

Base.Filesystem.rm — Method.

rm(rfs::RemoteFileSet, file; force=false)

Remove a specific downloaded file contained in rfs.

source

Base.Filesystem.rm — Method.

rm(rfs::RemoteFileSet; force=false)

Remove all downloaded files contained in rfs.

source

Base.Filesystem.rm — Method.

rm(rf::RemoteFile; force=false)

Remove the downloaded file rf.

source

Base.download — Method.

download(rfs::RemoteFileSet;
    quiet::Bool=false, verbose::Bool=false, force::Bool=false)

Download all files contained in rfs.

quiet: Do not print messages.

verbose: Print all messages.

force: Force download and overwrite existing files.

source

Base.download — Method.

download(rf::RemoteFile; quiet::Bool=false, verbose::Bool=false,
    force::Bool=false)

Download rf.

quiet: Do not print messages.

verbose: Print all messages.

force: Force download and overwrite existing files.

retries: Override the number of retries in rf

source

FileIO.load — Method.

load(rf::RemoteFile)

Load the contents of a remote file, downloading the file if it has not been done previously, reading the file from disk and trying to infer the format from filename and/or magic bytes in the file via FileIO.jl.

source

RemoteFiles.files — Method.

files(rfs::RemoteFileSet)

Get the (unsorted) list of file identifiers from a RemoteFileSet.

source

RemoteFiles.path — Method.

path(rfs::RemoteFileSet, file)

Get the path to a specific downloaded file contained in rfs.

source

RemoteFiles.path — Method.

path(rf::RemoteFile)

Get the local path of rf.

source

RemoteFiles.paths — Method.

paths(rfs::RemoteFileSet, files...)

Get the paths to specific downloaded files contained in rfs.

source

RemoteFiles.@RemoteFile — Macro.

@RemoteFile name url [key=value...]

Assign the RemoteFile located at url to the variable name.

The following keyword arguments are available:

file: Set a different local file name.

dir: The download directory. If dir is not set RemoteFiles will create a new directory data under the root of the current package and save the file there.

updates (default: :never): Indicates with which frequency the remote file is updated. Possible values are:

:never

:daily

:monthly

:yearly

:mondays/:weekly, :tuesdays, etc.

retries (default: 3): How many retries should be attempted.

wait (default: 5): How many seconds to wait between retries.

failed (default: :error): What to do if the download fails. Either throw an exception (:error) or display a warning (:warn).

source

RemoteFiles.@RemoteFileSet — Macro.

@RemoteFileSet name description begin
    file1 = @RemoteFile ...
    file2 = @RemoteFile ...
    ...
end

Collect several RemoteFiles in the RemoteFileSet saved under name with a description.

source